# API_BAJATEL/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bajatel - Proyecto DAW",
        "description": "API REST para gestión de contratos de telecomunicaciones (Fibra, TV, Móvil)",
        "contact": {
            "email": "soporte@bajatel.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/servicios/tv": {
            "post": {
                "tags": [
                    "Admin - TV"
                ],
                "summary": "Añadir opción de TV",
                "operationId": "8e562be4f57ab5b9fff7180d6b38deeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_paquete",
                                    "precio",
                                    "disponible"
                                ],
                                "properties": {
                                    "nombre_paquete": {
                                        "type": "string",
                                        "example": "Paquete Premium"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 19.99
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Opción añadida"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Requiere rol gestor/admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/servicios/tv/{id}": {
            "get": {
                "tags": [
                    "Admin - TV"
                ],
                "summary": "Mostrar opción de TV",
                "operationId": "9a23e1172a1cf96818e351bb0ecd1f48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opción encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - TV"
                ],
                "summary": "Editar opción de TV",
                "operationId": "5f45bb978536cfaa69dad4459c2d19a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_paquete": {
                                        "type": "string"
                                    },
                                    "precio": {
                                        "type": "number"
                                    },
                                    "disponible": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opción actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum - Usar: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}